requirements.txt
Django==1.6.7
42cc-pystyle
mock
django-extra-views==0.8
